requirements.txt
colorzero
gpiozero
lgpio
numpy
pillow
setuptools
smbus
spidev
